A tighter Delete protocol and change in return values.

+++ application/models/articles_model.php
@@ -75,15 +75,25 @@ class Articles_model extends CI_Model
 
 	public function delete($art_id)
 	{
-		$this->db->delete('articles', array('id' => $art_id));
-		if($this->db->affected_rows() >0)
+		$query_str = "SELECT user_lawyer.username FROM `articles` JOIN user_lawyer ON articles.user_id=user_lawyer.id WHERE articles.id='$art_id'";
+		$username = $this->db->query($query_str)->row_array()['username'];
+		if($username==$this->session->userdata('unnamed'))
 		{
-			return True;
+			$this->db->delete('articles', array('id' => $art_id));
+			if($this->db->affected_rows() >0)
+			{
+				return 1;
+			}
+			else
+			{
+				return 0;
+			}
 		}
 		else
 		{
-			return False;
+			return -1;
 		}
+
 	}
 
 	public function topic_list()
